Move classes from ``main.cc`` into includable file
``RGBMatrixManipulator`` and ``DisplayUpdater`` are both useful classes, but
having them in ``main.cc`` would have meant that users of these classes have to
reimplement them in their own code. Now they are located in ``led-matrix.h``,
where they can be included in client code more easily.

I left the implementation in the header because it's so small.