Given that there is now --led-cols, the use of -c is confusing. Deprecate.
In early versions of the libray, we used -r for --led-rows and -c for
--led-chain, but they were kept for the demo program to match old
documentation.

Now, there is also option --led-cols which makes the choice
of -c rather ambiguous (and the natural assumption of its meaning
being --led-cols wrong). So more actively message about the meaning
of the flag and discourage its use. Flag should be removed entirely
after a while.