Replace CanvasTransformer concept with PixelMapper.
The CanvasTransformer was meant to be more universal, but this complexity
was adding unnecesary complexity.

Introduce the new concept of a PixelMapper that with just the implementation
of a simple mapping function allows the same thing.
In this first step, the multiplex transformers are converted to the new scheme
and simplify them in the process. Only one place to add a multiplex transformer,
then the rest (command line support) works automatically.

Next steps:
  o replace the old UArrangementTransformer and RotateTransformer with
    the new scheme and only use these in all the demos and utilities.
  o Since PixelMappers now have a name, it is possible to implement a scheme
    by which they are registered and mentioned on the command line (for simple
    use in other language bindings).