Work around: mmap()-ing /dev/gpiomem is broken as it does not fail on unsupported offset.
Not only the PWM or 1Mhz counter regions are not supported, but /dev/gpiomem is
always failing silently: it always returns a memory region to GPIO but does not
verify its parameters.

Work around by only falling back to /dev/gpiomem if we request GPIO registers,
none of the others. Bail for the rest.

Filed https://github.com/raspberrypi/linux/issues/3055 to fix the broken implementation
in the kernel.