Add initialization for FM6126A panels. Use --led-panel-type=FM6126A
This is mostly experimental at this point (I only have a single panel
to test, thanks @esden), so please let me know if this works for you.
It is modelled after the Python script that was shared on #746.

Right now, this is not abstracted yet for potential other panel types,
this will be added as we go.

(This works in C++ now. If someone needs this using Python, please update
the Python bindings and send a pull request for the new option).

Fixes #746 #807