Make sure the rotating matrix behaves as before.
The way the rotate positions are set up is that they are offset by
cent_x in both directions, so we can't use the cent_x, cent_y when
offsetting.

This is confusing and should proably be replaced with a min_rotate
and max_rotate that are minus and plus rotate_square/2. Anyway fix
this for now that it behaves as before.

Update to pull #1148